Made the testing gatherer module, added it in utils directory and added name in author
